ITR data production checkpoint
Of the companies listed in the ITR V2 portfolio, 2 remain outside the analysis.  Both NEXTERA and PT&E have lingering LEI complexities that are really best addressed with better entity matching.  Otherwise, we have data for the remaining companies for which positions are established and for which SEC DERA data is available.