Update pipeline to work with 2023 RMI data
Many small adjustments so that this pipeline produces tables that can be used by the latest version of the Data Vault.  Perhaps the largest change was to remove emissions data from the target_data and trajectory_data tables we write.  Since we also write production_data and emissions_data, we have all the data we need, it's just not pre-joined.  But nothing made any use of such pre-joining, so we've dropped that for now.  Also, we now compute S1+S2 and S1+S2+S3 data, as well as supporting S3 data for Electricity Utilities.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>