Use production-centric rules always for Power Gen
The OECM scope rules for Power (aka Electricity) Utilities are just so different from corp disclosures that we really need to use production-centric rules for both PC and S3 cases.  And we need to let all-zero data (such as S3 data for Production-Centric Power Utilities) appear as zeros in the data, not the absence of data in the benchmark.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>